worked on suggested feedbacks
