--- client/src/components/AuthMenu/useStyles.ts
@@ -1,9 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-const useStyles = makeStyles(() => ({
-  profileLink: {
-    textDecoration: 'none',
-    color: '#000',
-  },
-}));
-
-export default useStyles;
