+++ client/src/components/Navbar/Navbar.tsx
@@ -1,13 +1,12 @@
 import { useState } from 'react';
 import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
-import Button from '@material-ui/core/Button';
 import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
 import useStyles from './useStyles';
-import { Link } from 'react-router-dom';
 import { useAuth } from '../../context/useAuthContext';
-import AuthMenu from '../AuthMenu/AuthMenu';
+import OtherNavLinks from './OtherNavLinks';
+import AuthNavLinks from './AuthNavLinks';
 
 const Navbar = (): JSX.Element => {
   const classes = useStyles();
@@ -17,7 +16,10 @@ const Navbar = (): JSX.Element => {
   return (
     <Grid container className={classes.root}>
       <Grid xs={5} md={3} className={classes.navbarBanner}>
-        <img src="https://res.cloudinary.com/dz8xmxmly/image/upload/v1636475121/fun_kickoff_project/logo_ddnmxs.png" />
+        <img
+          src="https://res.cloudinary.com/dz8xmxmly/image/upload/v1636475121/fun_kickoff_project/logo_ddnmxs.png"
+          alt="logo"
+        />
       </Grid>
       <Grid xs={6} item className={classes.menu}>
         <IconButton edge="start" color="inherit" onClick={() => setIsOpen(!isOpen)}>
@@ -30,38 +32,7 @@ const Navbar = (): JSX.Element => {
           flexDirection={{ xs: 'column', sm: 'row' }}
           className={classes.links}
         >
-          {!loggedInUser ? (
-            <>
-              <Link to="/login" className={classes.authNavLinks}>
-                <Button color="primary" variant="outlined">
-                  LOGIN
-                </Button>
-              </Link>
-              <Link to="/signup" className={classes.authNavLinks}>
-                <Button color="primary" variant="contained">
-                  SIGNUP
-                </Button>
-              </Link>
-            </>
-          ) : (
-            <>
-              <Button className={classes.navLinks} variant="text">
-                Become a sitter
-              </Button>
-              <Button className={classes.navLinks} variant="text">
-                Notifications
-              </Button>
-              <Button variant="text" className={classes.navLinks}>
-                <Link to="/bookings" className={classes.bookingLink}>
-                  Bookings
-                </Link>
-              </Button>
-              <Button className={classes.navLinks} variant="text">
-                Messages*
-              </Button>
-              <AuthMenu />
-            </>
-          )}
+          {!loggedInUser ? <AuthNavLinks /> : <OtherNavLinks />}
         </Box>
       </Grid>
     </Grid>
